+++ CreeDictionary/tests/DatabaseManager_tests/data/crkeng-pipon-of-different-word-classes/engcrk.xml
@@ -1,0 +1,1 @@
+<r/>
